backend/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
python-multipart>=0.0.5
pydantic>=1.8.0
python-dateutil>=2.8.2
requests>=2.26.0
llama-cpp-python>=0.2.0
tqdm>=4.66.0